% Auto-generated: do not edit by hand
\name{''HTMLAnnotator}

\alias{''HTMLAnnotator}

\title{HTMLAnnotator component}

\description{
HTMLAnnotator is used to annotate HTML. It renders an HTML string, which the user can highlight.
}

\usage{
''HTMLAnnotator(id=NULL, color=NULL, data=NULL, event=NULL, height=NULL,
highlights=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{color}{Character. Color of highlights.}

\item{data}{Character. HTML string.}

\item{event}{Named list. Name of event that has occurred.}

\item{height}{Character. Height of annotator container.}

\item{highlights}{Unnamed list. List of highlights applied to text.}

\item{width}{Character. Width of annotator container.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

